+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/compiler/common/Constants.java
@@ -232,6 +232,7 @@ public interface Constants {
     String METHOD_GET_TYPED_SUCCESSORS = "__getTypedSuccessors";
     String METHOD_GET_SUCCESSORS = "__getSuccessors";
     String METHOD_GET_RELATIONS = "__getRelations";
+    String METHOD_GET_BASE_RELATIONS = "getBaseRelations";
     String ARGS_2_BEAN = "args2bean";
     String GETTER_VAR = "getter";
     String ELEM_VAR = "elem";